+++ rollup.config.js
@@ -69,6 +69,33 @@ export default [
     context: 'window',
   },
   {
+    input: 'integration-tests/otel-api-globals.ts',
+    output: {
+      file: 'dist/artifacts/otel-api-globals.js',
+      format: 'iife',
+      name: 'OtelApiGlobals',
+      sourcemap: true,
+    },
+    plugins: [
+      json(),
+      nodeResolvePlugin,
+      commonjs({
+        include: /node_modules/,
+        sourceMap: true,
+        transformMixedEsModules: true,
+      }),
+      typescript({ tsconfig: './tsconfig.base.json' }),
+      babelPlugin,
+      terser({
+        ecma: 5,
+        output: {
+          comments: false
+        }
+      }),
+    ],
+    context: 'window',
+  },
+  {
     input: 'src/record/index.ts',
     output: {
       file: 'dist/artifacts/splunk-otel-web-record.js',